refactor: move `enforce` from `bin` to `services`
